requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
black==22.10.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
Django==4.1.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
executing==1.2.0
flake8==6.0.0
idna==3.4
ipdb==0.13.11
ipython==8.8.0
isort==5.10.1
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
Pygments==2.14.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
dj-rest-auth==2.1.11
django-allauth==0.48.0
djangorestframework-simplejwt==5.2.2
django-rest-authemail==2.1.4
